Add /maps to the v2.0 implementation
Still need to work out how to do the filtering allowed by /maps as
there's not really a connection between maps and datasets in Germinate.
