--- src/jhi/brapi/api/genomemaps/BrapiLinkageGroupMarker.java
@@ -1,26 +1,0 @@
-package jhi.brapi.api.genomemaps;
-
-public class BrapiLinkageGroupMarker
-{
-	private String markerDbId;
-	private String markerName;
-	private String location;
-
-	public String getMarkerDbId()
-		{ return markerDbId; }
-
-	public void setMarkerDbId(String markerDbId)
-		{ this.markerDbId = markerDbId; }
-
-	public String getMarkerName()
-		{ return markerName; }
-
-	public void setMarkerName(String markerName)
-		{ this.markerName = markerName; }
-
-	public String getLocation()
-		{ return location; }
-
-	public void setLocation(String location)
-		{ this.location = location; }
-}
